Groundwork for allowing code sharing between front and backend
This was complicated by the fact that it needed to work locally when using npm or when building production images in Docker. The solution isn't pretty but does work well with 2 docker files + docker ignore files at the root and using the DOCKER_BUILDKIT environment variable to add both the front/back-ends and the shared folder to the Docker build context.

Updated message box font to match the rest of the app now in lovely Verdana

Added CORS for cross-domain requests so we can run locally

Deployment workflow now only triggers from pushes to master so branches should be safer to create