Add all provinces when starting and populate your starting province with the correct stuff
Haven is now populated with the correct walls, units, roads and structures however the tile offsets are incorrect so everything shoes as a 3x3 hoverpad. Only requires updating the structures.json file but this is a job for another day...

This now breaks the overview map screen. I really need to figure out how to extract the provinces graphic so you can properly visualise connected provinces and their owners